Assert callback can create custom response
After the SAML authentication is concluded, the application needs to be free to create a custom response. Moreover, if you just redirect, cookies won't be set (Chrome).
Now the reply object is given to the callback as an extra argument, and it should always be returned to the caller too.